+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Updated dependencies.
 - Added support for `Vec<u8>` data type (sqlite `BLOB`).
 - Renamed `try_insert` to `insert` and `insert` to `insert_ok`.
+- Renamed `try_delete` to `delete` and `delete` to `delete_ok`.
 
 # 0.3.1